Add Body entries
- Add BodyEntry
- Further entry refactoring
