+++ MicroEngineerProject/MicroEngineer/Entries/FlightEntries.cs
@@ -127,7 +127,7 @@ namespace MicroMod
             Description = "Shows the ratio of total lift and drag forces.";
             Category = MicroEntryCategory.Flight;
             Unit = null;
-            Formatting = "{0:N3}";
+            Formatting = "{0:N2}";
         }
 
         public override void RefreshData()
@@ -185,4 +185,177 @@ namespace MicroMod
 
         public override string ValueDisplay => base.ValueDisplay;
     }
+
+    public class DragCoefficient : FlightEntry
+    {
+        public DragCoefficient()
+        {
+            Name = "DragCoefficient";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = null;
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.DragCoefficient;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class ExposedArea : FlightEntry
+    {
+        public ExposedArea()
+        {
+            Name = "ExposedArea";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = null; // TODO
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.ExposedArea;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class Heading : FlightEntry
+    {
+        public Heading()
+        {
+            Name = "Heading";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Heading;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class Pitch_HorizonRelative : FlightEntry
+    {
+        public Pitch_HorizonRelative()
+        {
+            Name = "Pitch";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Pitch_HorizonRelative;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class Roll_HorizonRelative : FlightEntry
+    {
+        public Roll_HorizonRelative()
+        {
+            Name = "Roll";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Roll_HorizonRelative;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class Yaw_HorizonRelative : FlightEntry
+    {
+        public Yaw_HorizonRelative()
+        {
+            Name = "Yaw";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Yaw_HorizonRelative;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class Zenith : FlightEntry
+    {
+        public Zenith()
+        {
+            Name = "Zenith";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Zenith;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class SoundSpeed : FlightEntry
+    {
+        public SoundSpeed()
+        {
+            Name = "Speed of sound";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "m/s";
+            Formatting = "{0:N1}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.SoundSpeed;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class GeeForce : FlightEntry
+    {
+        public GeeForce()
+        {
+            Name = "G-Force";
+            Description = "";
+            Category = MicroEntryCategory.Flight;
+            Unit = "g";
+            Formatting = "{0:N3}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.geeForce;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+
 }
