+++ MicroEngineerProject/MicroEngineer/Entries/OrbitalEntries.cs
@@ -193,4 +193,287 @@ namespace MicroMod
             }
         }
     }
+
+    public class OrbitalSpeed : OrbitalEntry
+    {
+        public OrbitalSpeed()
+        {
+            Name = "Orbital Velocity";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "m/s";
+            Formatting = "{0:N1}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.OrbitalSpeed;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class EccentricAnomaly : OrbitalEntry
+    {
+        public EccentricAnomaly()
+        {
+            Name = "Eccentric Anomaly";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.EccentricAnomaly;
+        }
+
+        public override string ValueDisplay
+        {
+            get
+            {
+                if (EntryValue == null)
+                    return "-";
+
+                return String.IsNullOrEmpty(base.Formatting) ? EntryValue.ToString() : String.Format(Formatting, (double)EntryValue * PatchedConicsOrbit.Rad2Deg);
+            }
+        }
+    }
+
+    public class MeanAnomaly : OrbitalEntry
+    {
+        public MeanAnomaly()
+        {
+            Name = "Mean Anomaly";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.MeanAnomaly;
+        }
+
+        public override string ValueDisplay
+        {
+            get
+            {
+                if (EntryValue == null)
+                    return "-";
+
+                return String.IsNullOrEmpty(base.Formatting) ? EntryValue.ToString() : String.Format(Formatting, (double)EntryValue * PatchedConicsOrbit.Rad2Deg);
+            }
+        }
+    }
+
+    public class TrueAnomaly : OrbitalEntry
+    {
+        public TrueAnomaly()
+        {
+            Name = "True Anomaly";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "°";
+            Formatting = "{0:N1}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.TrueAnomaly;
+        }
+
+        public override string ValueDisplay
+        {
+            get
+            {
+                if (EntryValue == null)
+                    return "-";
+
+                return String.IsNullOrEmpty(base.Formatting) ? EntryValue.ToString() : String.Format(Formatting, MicroUtility.RadiansToDegrees((double)EntryValue));
+            }
+        }
+    }
+
+    public class ObT : OrbitalEntry
+    {
+        public ObT()
+        {
+            Name = "Orbit Time";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "s";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.ObT;
+        }
+
+        public override string ValueDisplay
+        {
+            get
+            {
+                if (EntryValue == null)
+                    return "-";
+
+                return String.IsNullOrEmpty(base.Formatting) ? EntryValue.ToString() : String.Format(Formatting, MicroUtility.SecondsToTimeString((double)EntryValue, true, false));
+            }
+        }
+    }
+
+    public class ArgumentOfPeriapsis : OrbitalEntry
+    {
+        public ArgumentOfPeriapsis()
+        {
+            Name = "Argument of Pe. ω";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.OrbitalElements.ArgumentOfPeriapsis;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class LongitudeOfAscendingNode : OrbitalEntry
+    {
+        public LongitudeOfAscendingNode()
+        {
+            Name = "LAN Ω";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "°";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.OrbitalElements.LongitudeOfAscendingNode;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class SemiMajorAxis : OrbitalEntry
+    {
+        public SemiMajorAxis()
+        {
+            Name = "Semi Major Axis <i>a</i>";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "m";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.OrbitalElements.SemiMajorAxis;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class SemiMinorAxis : OrbitalEntry
+    {
+        public SemiMinorAxis()
+        {
+            Name = "Semi Minor Axis <i>b</i>";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "m";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.SemiMinorAxis;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class OrbitalEnergy : OrbitalEntry
+    {
+        public OrbitalEnergy()
+        {
+            Name = "Orbital Energy";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "kJ";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.OrbitalEnergy / 1000.0;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class SemiLatusRectum : OrbitalEntry
+    {
+        public SemiLatusRectum()
+        {
+            Name = "Semi Latus Rectum ℓ";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "m";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.SemiLatusRectum;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class OrbitPercent : OrbitalEntry
+    {
+        public OrbitPercent()
+        {
+            Name = "Orbit Percent";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "%";
+            Formatting = "{0:N2}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.orbitPercent * 100;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
+
+    public class OrbitRadius : OrbitalEntry
+    {
+        public OrbitRadius()
+        {
+            Name = "Orbit Radius";
+            Description = "";
+            Category = MicroEntryCategory.Orbital;
+            Unit = "m";
+            Formatting = "{0:N0}";
+        }
+
+        public override void RefreshData()
+        {
+            EntryValue = MicroUtility.ActiveVessel.Orbit.radius;
+        }
+
+        public override string ValueDisplay => base.ValueDisplay;
+    }
 }
